Budget api

+++ backend/underbudget/models/envelope.py
@@ -45,6 +45,8 @@ class EnvelopeModel(db.Model, AuditModel, CrudModel):
         db.Integer, db.ForeignKey("envelope_category.id"), nullable=False
     )
     transactions = db.relationship("EnvelopeTransactionModel", lazy="select")
+    periodic_expenses = db.relationship("BudgetPeriodicExpenseModel", lazy="select")
+    annual_expenses = db.relationship("BudgetAnnualExpenseModel", lazy="select")
 
     name = db.Column(db.String(128), nullable=False)
     archived = db.Column(db.Boolean, nullable=False)
@@ -54,4 +56,8 @@ class EnvelopeModel(db.Model, AuditModel, CrudModel):
         """ Deletes the envelope if it does not have any associated transactions """
         if len(self.transactions) > 0:
             raise Conflict("Envelope is referenced by transactions")
+        if len(self.periodic_expenses) > 0:
+            raise Conflict("Envelope is referenced by budgeted periodic expenses")
+        if len(self.annual_expenses) > 0:
+            raise Conflict("Envelope is referenced by budgeted annual expenses")
         super().delete()
